requirements.txt
requests
beautifulsoup4
pandas
tqdm
sentence-transformers
faiss-cpu